requirements.txt
flask
python-dotenv
anthropic
bcrypt
psycopg2-binary
gunicorn
postgresql 
postgresql-contrib